kissanai_improved/backend/requirements.txt
Flask 
Flask-Session
SpeechRecognition
gTTS
googletrans
python-dotenv